+++ frontend/src/components/Diagram/LineChartBlock.js
