+++ frontend/src/components/Diagram/ScatterPlotBlock.js
