+++ frontend/src/components/index.js
@@ -1,3 +1,4 @@
+import Diagram from './Diagram';
 import Container from './Container';
 import Logo from './Logo';
 import Column from './Column';
@@ -5,4 +6,4 @@ import Row from './Row';
 import FlexDiv from './FlexDiv';
 import Selector from './Selector';
 
-export { Container, Logo, Column, Row, FlexDiv, Selector };
+export { Diagram, Container, Logo, Column, Row, FlexDiv, Selector };
